@@ -1,16 +1,15 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-
-import { Auth } from "./hooks/auth";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(
+import App from "./App";
+import { Auth } from "./hooks/auth";
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
   <Auth>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </Auth>,
-  document.getElementById("root")
+  </Auth>
 );
