@@ -1,28 +1,53 @@
+import aquaticEnemy from './aquatic_enemy.png';
+import aquaticEnemyhover from './aquatic_enemyhover.png';
+import aquaticFriendly from './aquatic_friendly.png';
+import aquaticFriendlyhover from './aquatic_friendlyhover.png';
+import flyingEnemy from './flying_enemy.png';
+import flyingEnemyhover from './flying_enemyhover.png';
+import flyingFriendly from './flying_friendly.png';
+import flyingFriendlyhover from './flying_friendlyhover.png';
+import jumpEnemy from './jump_enemy.png';
+import jumpEnemyhover from './jump_enemyhover.png';
+import jumpFriendly from './jump_friendly.png';
+import jumpFriendlyhover from './jump_friendlyhover.png';
+import rangedEnemy from './ranged_enemy.png';
+import rangedEnemyhover from './ranged_enemyhover.png';
+import rangedFriendly from './ranged_friendly.png';
+import rangedFriendlyhover from './ranged_friendlyhover.png';
+import scootEnemy from './scoot_enemy.png';
+import scootEnemyhover from './scoot_enemyhover.png';
+import scootFriendly from './scoot_friendly.png';
+import scootFriendlyhover from './scoot_friendlyhover.png';
+import tauntEnemy from './taunt_enemy.png';
+import tauntEnemyhover from './taunt_enemyhover.png';
+import tauntFriendly from './taunt_friendly.png';
+import tauntFriendlyhover from './taunt_friendlyhover.png';
+
 const icons = {
-    aquatic_enemy: 'assets/images/ui/aquatic_enemy.png',
-    aquatic_enemyhover: 'assets/images/ui/aquatic_enemyhover.png',
-    aquatic_friendly: 'assets/images/ui/aquatic_friendly.png',
-    aquatic_friendlyhover: 'assets/images/ui/aquatic_friendlyhover.png',
-    flying_enemy: 'assets/images/ui/flying_enemy.png',
-    flying_enemyhover: 'assets/images/ui/flying_enemyhover.png',
-    flying_friendly: 'assets/images/ui/flying_friendly.png',
-    flying_friendlyhover: 'assets/images/ui/flying_friendlyhover.png',
-    jump_enemy: 'assets/images/ui/jump_enemy.png',
-    jump_enemyhover: 'assets/images/ui/jump_enemyhover.png',
-    jump_friendly: 'assets/images/ui/jump_friendly.png',
-    jump_friendlyhover: 'assets/images/ui/jump_friendlyhover.png',
-    ranged_enemy: 'assets/images/ui/ranged_enemy.png',
-    ranged_enemyhover: 'assets/images/ui/ranged_enemyhover.png',
-    ranged_friendly: 'assets/images/ui/ranged_friendly.png',
-    ranged_friendlyhover: 'assets/images/ui/ranged_friendlyhover.png',
-    scoot_enemy: 'assets/images/ui/scoot_enemy.png',
-    scoot_enemyhover: 'assets/images/ui/scoot_enemyhover.png',
-    scoot_friendly: 'assets/images/ui/scoot_friendly.png',
-    scoot_friendlyhover: 'assets/images/ui/scoot_friendlyhover.png',
-    taunt_enemy: 'assets/images/ui/taunt_enemy.png',
-    taunt_enemyhover: 'assets/images/ui/taunt_enemyhover.png',
-    taunt_friendly: 'assets/images/ui/taunt_friendly.png',
-    taunt_friendlyhover: 'assets/images/ui/taunt_friendlyhover.png',
+    aquatic_enemy: aquaticEnemy,
+    aquatic_enemyhover: aquaticEnemyhover,
+    aquatic_friendly: aquaticFriendly,
+    aquatic_friendlyhover: aquaticFriendlyhover,
+    flying_enemy: flyingEnemy,
+    flying_enemyhover: flyingEnemyhover,
+    flying_friendly: flyingFriendly,
+    flying_friendlyhover: flyingFriendlyhover,
+    jump_enemy: jumpEnemy,
+    jump_enemyhover: jumpEnemyhover,
+    jump_friendly: jumpFriendly,
+    jump_friendlyhover: jumpFriendlyhover,
+    ranged_enemy: rangedEnemy,
+    ranged_enemyhover: rangedEnemyhover,
+    ranged_friendly: rangedFriendly,
+    ranged_friendlyhover: rangedFriendlyhover,
+    scoot_enemy: scootEnemy,
+    scoot_enemyhover: scootEnemyhover,
+    scoot_friendly: scootFriendly,
+    scoot_friendlyhover: scootFriendlyhover,
+    taunt_enemy: tauntEnemy,
+    taunt_enemyhover: tauntEnemyhover,
+    taunt_friendly: tauntFriendly,
+    taunt_friendlyhover: tauntFriendlyhover,
 }
 
 export default icons;
