@@ -1,6 +1,11 @@
+import plusCheekia from "./plus_cheekia.png";
+import plusCard from "./plus_card.png";
+import cheekWinner from "./cheek_winner.png";
+
 const images_ui = {
-    'wheel-B3': 'assets/images/ui/pluscard.png',
-    'wheel-C2': 'assets/images/ui/pluscheekia.png',
-}
+  "wheel-B3": plusCard,
+  "wheel-C2": plusCheekia,
+  "cheekWinner": cheekWinner
+};
 
 export default images_ui;
