@@ -3,7 +3,7 @@ import "./ChooseModal.css";
 import Card from "../../Game/Hand/Card/Card";
 import { useStore } from "../../../../hooks/store";
 
-const ChooseModal = (props) => {
+const ChooseModal = React.memo((props) => {
   const [state, dispatch] = useStore();
 
   const chooseHandler = (event) => {
@@ -35,12 +35,12 @@ const ChooseModal = (props) => {
   });
 
   return (
-    <div className="modal">
+    <div className="choose-modal">
       <div style={{ justifyContent: "center", display: "flex", height: "75%" }}>
         {cards}
       </div>
     </div>
   );
-};
+});
 
 export default ChooseModal;
